Split test-pg-FastAPI/src/main.py Fixes #3
